Infer end position for Pylint diagnostics
Pylint only supplies the start position of markers, but not the end.
Always extending the marker until the end of the line leads to
undesirable highlighting in editors, for example:

```python
def my_func(unused_arg, used_arg):
    print(used_arg)
```

Starting an "unused argument" diagnostic at `unused_arg` and extending
It until EOL would also highlight `used_arg`, even though the diagnostic
doesn't apply there.

With this commit, the end position sent to the editor is automatically
determined by looking for an identifier at the indicated start position.
If there is no valid identifier at that position, the entire line is
highlighted as before.